pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel", "cython", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pyranges_plot"
version = "0.1.0"
description = "Genomic intervals data visualization package for dataframe objects generated with PyRanges."
requires-python = ">=3.12"
readme = "README.md"
authors = [
    { name = "Ester Muñoz del Campo", email = "ester.munoz01@estudiant.upf.edu" },
    { name = "Max Ticó Miñarro", email = "max.tico@ub.edu"},
    { name = "Marco Mariotti", email = "marco.mariotti@ub.edu" },
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["bioinformatics", "genomicranges", "genomics", "graphics"]
dependencies = ["pyranges1>=1.0.0", "intervaltree>=3.1.0", "kaleido>=0.2.1"]

[project.optional-dependencies]
matplotlib = [
    "matplotlib >= 3.7.2",
    "mplcursors >= 0.5.2",
]

plotly = [
    "plotly >= 5.9.0",
    "dash >= 2.14.0",
    "dash_bootstrap_components >= 1.5.0",
]
dev = [
    "pytest",
    "pytest-mpl",
    "tox", 
    "ruff == 0.3.0",
    "pillow",
]
docs = [
    "sphinx >= 7.3.7",
    "sphinx_rtd_theme >= 2.0.0",
    "sphinx-autoapi >= 3.1.1",
    "sphinxcontrib-napoleon >= 0.7",
    "memory-profiler >= 0.61.0",
]
all = [
    "matplotlib >= 3.7.2",
    "mplcursors >= 0.5.2",
    "plotly >= 5.9.0",
    "dash >= 2.14.0",
    "dash_bootstrap_components >= 1.5.0",
    "sphinx >= 7.3.7",
    "sphinx_rtd_theme >= 2.0.0",
    "sphinx-autoapi >= 3.1.1",
    "sphinxcontrib-napoleon >= 0.7",
    "memory-profiler >= 0.61.0",
]

[project.urls]
Homepage="https://github.com/pyranges/pyranges_plot"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py312

[testenv]
deps =
    pytest
    pytest-mpl
    tox
    ruff == 0.3.0
    pillow
    plotly >= 5.9.0
    dash >= 2.14.0
    dash_bootstrap_components >= 1.5.0
    matplotlib >= 3.7.2
    mplcursors >= 0.5.2
    pyranges1>=1.0.0
    intervaltree>=3.1.0
    kaleido>=0.2.1
commands = 
    pytest tests/test_ply.py
    pytest tests/test_mpl.py --mpl --mpl-results-path=results --mpl-generate-summary=html
"""
